py/runtime: __int__: Restore "can't convert <type> to int" error message.
After implementing support for the __int__ special method, error when
passing an incompatible type where int is expected changed from:

    can't convert <type> to int

to:

    unsupported type for __int__: '<type>'

It's anticipated that the latter error message may be confusing for people
not familiar with Python intrinsics like special methods. So, restore
the older error message specifically for __int__ operation in mp_unary_op().